+++ Cargo.toml
@@ -1,150 +1,8 @@
 [profile.release]
-panic = "unwind"
-
-[package]
-name = "subsocial-node"
-version = "2.0.0"
-build = "build.rs"
-authors = ["DappForce <dappforce@pm.me>"]
-edition = "2018"
-license = "GPL-3.0-only"
-homepage = "https://subsocial.network"
-repository = "https://github.com/dappforce/dappforce-subsocial-node"
-description = "Reference implementation of Subsocial node"
-keywords = ["blockchain", "cryptocurrency", "social-network", "news-feed", "marketplace"]
-categories = ["cryptography::cryptocurrencies"]
-
-[build-dependencies]
-vergen = "3.0.4"
-
-[build-dependencies.build-script-utils]
-git = "https://github.com/paritytech/substrate.git"
-package = "substrate-build-script-utils"
-rev = "3e651110aa06aa835790df63410a29676243fc54"
-version = "2.0.0"
-
-[dependencies]
-futures = "0.3.1"
-log = "0.4.8"
-parking_lot = "0.9.0"
-tokio = "0.1.22"
-trie-root = "0.15.2"
-hex-literal = "0.2.1"
-
-[dependencies.codec]
-package = "parity-scale-codec"
-version = "1.0.0"
-
-[dependencies.ctrlc]
-features = ["termination"]
-version = "3.1.3"
-
-[dependencies.futures01]
-package = "futures"
-version = "0.1.29"
-
-[dependencies.grandpa]
-git = "https://github.com/paritytech/substrate.git"
-package = "sc-finality-grandpa"
-rev = "3e651110aa06aa835790df63410a29676243fc54"
-version = "2.0.0"
-
-[dependencies.grandpa-primitives]
-git = "https://github.com/paritytech/substrate.git"
-package = "sp-finality-grandpa"
-rev = "3e651110aa06aa835790df63410a29676243fc54"
-version = "2.0.0"
-
-[dependencies.subsocial-runtime]
-path = "runtime"
-version = "2.0.0"
-
-[dependencies.sc-basic-authority]
-git = "https://github.com/paritytech/substrate.git"
-rev = "3e651110aa06aa835790df63410a29676243fc54"
-
-[dependencies.sc-cli]
-git = "https://github.com/paritytech/substrate.git"
-rev = "3e651110aa06aa835790df63410a29676243fc54"
-version = "2.0.0"
-
-[dependencies.sc-client]
-git = "https://github.com/paritytech/substrate.git"
-rev = "3e651110aa06aa835790df63410a29676243fc54"
-version = "2.0.0"
-
-[dependencies.sc-consensus-aura]
-git = "https://github.com/paritytech/substrate.git"
-rev = "3e651110aa06aa835790df63410a29676243fc54"
-version = "0.8"
-
-[dependencies.sc-executor]
-git = "https://github.com/paritytech/substrate.git"
-rev = "3e651110aa06aa835790df63410a29676243fc54"
-version = "2.0.0"
-
-[dependencies.sc-network]
-git = "https://github.com/paritytech/substrate.git"
-rev = "3e651110aa06aa835790df63410a29676243fc54"
-version = "0.8"
-
-[dependencies.sc-service]
-git = "https://github.com/paritytech/substrate.git"
-rev = "3e651110aa06aa835790df63410a29676243fc54"
-version = "2.0.0"
-
-[dependencies.sc-transaction-pool]
-git = "https://github.com/paritytech/substrate.git"
-rev = "3e651110aa06aa835790df63410a29676243fc54"
-version = "2.0.0"
-
-[dependencies.telemetry]
-package = "sc-telemetry"
-git = "https://github.com/paritytech/substrate.git"
-rev = "3e651110aa06aa835790df63410a29676243fc54"
-version = "2.0.0"
-
-[dependencies.sp-consensus]
-git = "https://github.com/paritytech/substrate.git"
-rev = "3e651110aa06aa835790df63410a29676243fc54"
-version = "0.8"
-
-[dependencies.sp-consensus-aura]
-git = "https://github.com/paritytech/substrate.git"
-rev = "3e651110aa06aa835790df63410a29676243fc54"
-version = "0.8"
-
-[dependencies.sp-core]
-git = "https://github.com/paritytech/substrate.git"
-rev = "3e651110aa06aa835790df63410a29676243fc54"
-version = "2.0.0"
-
-[dependencies.sp-inherents]
-git = "https://github.com/paritytech/substrate.git"
-rev = "3e651110aa06aa835790df63410a29676243fc54"
-version = "2.0.0"
-
-[dependencies.sp-io]
-git = "https://github.com/paritytech/substrate.git"
-rev = "3e651110aa06aa835790df63410a29676243fc54"
-version = "2.0.0"
-
-[dependencies.sp-runtime]
-git = "https://github.com/paritytech/substrate.git"
-rev = "3e651110aa06aa835790df63410a29676243fc54"
-version = "2.0.0"
-
-[dependencies.sp-transaction-pool]
-git = "https://github.com/paritytech/substrate.git"
-rev = "3e651110aa06aa835790df63410a29676243fc54"
-version = "2.0.0"
+panic = 'unwind'
 
 [workspace]
 members = [
-    "runtime",
-    "pallets/df-integration-tests"
+    'node',
+    'runtime',
 ]
-
-[[bin]]
-name = "subsocial-node"
-path = "src/main.rs"
