+++ node/src/cli.rs
@@ -1,0 +1,11 @@
+use sc_cli::{RunCmd, Subcommand};
+use structopt::StructOpt;
+
+#[derive(Debug, StructOpt)]
+pub struct Cli {
+	#[structopt(subcommand)]
+	pub subcommand: Option<Subcommand>,
+
+	#[structopt(flatten)]
+	pub run: RunCmd,
+}
